==> code/1c-pull_coinmetrics.ipynb <==
import pandas as pd
import numpy as np
from coinmetrics.api_client import CoinMetricsClient
import time
import pickle
import requests
from datetime import datetime
import time
from typing import Dict, List
import logging
import pytz

def makeCmApiCall(base_url: str, endpoint: str, params: Dict[str, str], num_retries: int = 3) -> requests.Response:
    """
    Makes an API call to the given endpoint with the given parameters.

    Args:
    - base_url (str): string representing the base URL for the API.
    - endpoint (str): string representing the endpoint to call.
    - params (Dict): dictionary containing the parameters for the API call.
    - num_retries (int): integer representing the number of times to retry the API call in case of an error.

    Returns:
    - response (requests.Response): the response object from the API call.
    """
    url = f"{base_url}/{endpoint}"
    retries = 0
    while retries < num_retries:
        try:
            response = requests.get(url, params=params, timeout=10)
            if response.status_code == 403:
                try:
                    return params['metrics']
                except:
                    raise Exception(f"Failed to make API call with response {response.status_code}.")
            response.raise_for_status()
            return response
        except requests.exceptions.Timeout as e:
            logging.warning(f"Timeout error occurred: {e}")
        except requests.exceptions.RequestException as e:
            logging.warning(f"Request error occurred: {e}")
        retries += 1
        if retries == 1:
            time.sleep(1)
        elif retries == 2:
            time.sleep(20)
    raise Exception(f"Failed to make API call after {num_retries} retries")

def formCoinmetricsAssetUniverse(client: CoinMetricsClient, cmc_assets_fp: str) -> pd.DataFrame:
    """ map cmc universe to coinmetrics universe. 
    (1) pull all cm assets and open my universe of cmc assets.
    (2) adjust cm asset names so they match to my cmc assets.
    (3) merge asset ids together on both the cm asset id and the full name.
    (4) clean the merged data.
    (5) add assets from cm that should be in the universe but aren't in cmc.
    (6) remove stablecoins and derivatives.

    Args:
        client (CoinMetricsClient): cm client object for pinging api.
        cmc_assets_fp (str): filepath to cmc asset universe pickle.
    
    Returns:
        merged_df (pd.DataFrame): dataframe of crosswalk between cmc id and cm id.    
    """
    # import cmc token universe
    with open(cmc_assets_fp, 'rb') as f:
        cmc_asset_universe_dict = pickle.load(f)

    # form unique cmc asset df
    cmc_assets = []
    for k, v in cmc_asset_universe_dict.items():
        cmc_assets.extend(v)
    cmc_assets = list(np.unique(np.array(cmc_assets)))
    cmc_assets_df = pd.DataFrame(data={'asset_cmc': cmc_assets})

    # pull all cm assets
    full_asset_catalog = client.catalog_full_assets()
    cm_assets_df = pd.DataFrame(full_asset_catalog)

    # Check that the "asset" column is unique in both dataframes
    assert (cmc_assets_df["asset_cmc"].is_unique 
            and cm_assets_df["full_name"].is_unique 
            and cm_assets_df['asset'].is_unique)

    # remove duplicated cm asset; they have a data error
    cm_assets_df = cm_assets_df[~cm_assets_df.asset.isin(['seed', 'tree', 'aurora'])]

    # change cm full names before merge so they match cmc for known nonmatches
    cm_assets_df.loc[cm_assets_df.asset=='aave', 'full_name'] = 'aave-old'
    cm_assets_df.loc[cm_assets_df.asset=='alpha', 'full_name'] = 'alpha-finance-lab'
    cm_assets_df.loc[cm_assets_df.asset=='mco', 'full_name'] = 'crypto-com'
    cm_assets_df.loc[cm_assets_df.asset=='fet', 'full_name'] = 'fetch'
    cm_assets_df.loc[cm_assets_df.asset=='clv', 'full_name'] = 'clover'
    cm_assets_df.loc[cm_assets_df.asset=='gno', 'full_name'] = 'gnosis-gno'
    cm_assets_df.loc[cm_assets_df.asset=='glm', 'full_name'] = 'golem-network-tokens'
    cm_assets_df.loc[cm_assets_df.asset=='hive', 'full_name'] = 'hive-blockchain'
    cm_assets_df.loc[cm_assets_df.asset=='rook', 'full_name'] = 'keeperdao'
    cm_assets_df.loc[cm_assets_df.asset=='yffii', 'full_name'] = 'yearn-finance-ii'
    cm_assets_df.loc[cm_assets_df.asset=='btt', 'full_name'] = 'bittorrent'
    cm_assets_df.loc[cm_assets_df.asset=='idex', 'full_name'] = 'aurora'
    cm_assets_df.loc[cm_assets_df.asset=='egld', 'full_name'] = 'multiversx-egld'
    cm_assets_df.loc[cm_assets_df.asset=='cfx', 'full_name'] = 'confluxnetwork'
    cm_assets_df.loc[cm_assets_df.asset=='xch', 'full_name'] = 'chia-network'
    cm_assets_df.loc[cm_assets_df.asset=='syn', 'full_name'] = 'synapse2'

    # clean the asset names to just low case letters and numbers and merge
    cmc_assets_df["asset_clean"] = cmc_assets_df["asset_cmc"].str.lower().str.replace(r"[^a-zA-Z0-9]", "")
    cm_assets_df["asset_clean"] = cm_assets_df["full_name"].str.lower().str.replace(r"[^a-zA-Z0-9]", "")
    merged_df = pd.merge(cmc_assets_df, cm_assets_df, 
                        on="asset_clean", how='inner',
                        validate='one_to_one')

    # repeat but use the unique asset abbreviation id from cm
    cm_assets_df["asset_clean"] = cm_assets_df["asset"].str.lower().str.replace(r"[^a-zA-Z0-9]", "")
    merged_df2 = pd.merge(cmc_assets_df, cm_assets_df, 
                        on="asset_clean", how='inner',
                        validate='one_to_one')

    # remove duplicated assets from the two merged dataframes and put them together
    merged_df2 = merged_df2[~merged_df2.asset.isin(list(merged_df.asset.values))]
    merged_df = pd.concat((merged_df, merged_df2))
    assert merged_df.asset.is_unique

    # clean up the merged data
    merged_df = merged_df[['asset_cmc', 'asset']]
    merged_df = merged_df.rename(columns={'asset': 'asset_cm'})
    merged_df = merged_df.reset_index(drop=True)

    # manually add to my universe of cm assets these assets to consider
    assets_to_add = ['ape', 'apt', 'arpa', 'badger', 'bal', 'cake', 'cel', 'comp', 'cvx',
        'dot', 'etc', 'fil', 'flr', 'flux', 'ftt', 'fun', 'gmx', 'grin', 'hnt',
        'inv', 'knc', 'krl', 'luna', 'luna2', 'mir', 'multi', 'nft', 'nu', 'ocean', 'ohm',
        'op', 'poly', 'qi', 'rndr', 'rpl', 'skl', 'snt', 'theta', 'tru', 'xdc', 'zrx']
    merged_df = pd.concat((merged_df, pd.DataFrame(data={'asset_cmc': np.repeat(np.nan, len(assets_to_add)),
                                                        'asset_cm': assets_to_add})))

    # manually remove stables and derivatives
    merged_df = merged_df[~merged_df.asset_cm.isin(['steth', 'wbtc', 'tusd', 'gusd', 'usdd', 'btcb'])]

    # manually add in both aave old and new
    merged_df = pd.concat((merged_df, pd.DataFrame(data={'asset_cmc': ['aave'],
                                                        'asset_cm': ['aave']}))).reset_index(drop=True)

    return merged_df


def pullAssetMetrics(client: CoinMetricsClient, base_url: str, base_params: dict, 
    cm_asset_universe: list, target_asset_metrics: list, num_retries=3):
    """
    Pulls asset metrics for assets in the given asset universe using the given base URL and parameters.

    Args:
    - client (CoinMetricsClient): client object for interacting with the CM API.
    - base_url: string representing the base URL for the API.
    - base_params: dictionary containing the base parameters for the API call.
    - cm_asset_universe: list of strings representing the assets to pull metrics for.
    - target_asset_metrics: list of strings of asset metrics of interest.
    - num_retries: integer representing the number of times to retry the API call in case of an error.

    Returns:
    - results_df: Pandas DataFrame containing the asset metrics.
    """

    # Initialize dataframe to return results
    results_df = pd.DataFrame(data={'asset': [], 'time': []})

    # Pull all the metrics for all assets
    asset_metrics_df = pd.DataFrame(client.catalog_full_assets())

    # Cut down to CM universe
    asset_metrics_df = asset_metrics_df[asset_metrics_df.asset.isin(cm_asset_universe)]
    assert asset_metrics_df.asset.is_unique
    assert asset_metrics_df.shape[0] == cm_asset_universe
    
    # Define endpoint and parameters
    endpoint = 'timeseries/asset-metrics'
    params = base_params.copy()
    params['page_size'] = 10000

    # For every asset, pull all metrics available for this asset
    for i in range(len(cm_asset_universe)):
        # Update asset
        asset = cm_asset_universe[i]

        # Monitor progress
        print(f"Processing the {i+1}th asset ({(i+1)/len(cm_asset_universe)*100:.2f}%): {asset}")

        # Initialize object for this asset results
        asset_results_df = pd.DataFrame(data={'asset': [], 'time': []})

        # update parameters for this asset
        params['assets'] = asset

        # determine all metrics for this asset (skip if there aren't metrics for this asset)
        metrics = asset_metrics_df[asset_metrics_df.asset==asset].metrics.values[0]
        if type(metrics) is not list:
            continue
        metrics_df = pd.DataFrame(metrics)

        # pull data for each metric
        assert metrics_df.metric.is_unique
        for metric in list(metrics_df.metric.values):
            # Skip if not a metric of interest
            if metric not in target_asset_metrics:
                continue

            # Print out the metric we are doing
            print(f"Working on metric: {metric}.")

            # form dataframe of the different frequency options for this metric
            metric_options_df = pd.DataFrame(metrics_df[metrics_df.metric==metric].frequencies.values[0])

            # Set frequency to 1d but report if it does not have a one day option and stop execution
            if '1d' in list(metric_options_df.frequency.values):
                frequency = '1d'
            else:
                print(metric_options_df)
                print(f"The metric {metric} for asset {asset} does not have a 1d frequency option.")
                continue
        
            # Set the start and end time for this frequency
            start_time = metric_options_df[metric_options_df.frequency==frequency].min_time.values[0]
            end_time = metric_options_df[metric_options_df.frequency==frequency].max_time.values[0]

            # Update params for this metric and frequency
            params['metrics'] = metric
            params['frequency'] = frequency
            params['start_time'] = start_time
            params['end_time'] = end_time

            # Make the API request
            response = makeCmApiCall(base_url, endpoint, params)
            if type(response)==str:
                print(f'The metric {metric} is not available for my API key.')
                continue
            data = response.json()['data']
            asset_df = pd.DataFrame(data)

            # Confirm we obtained the expected number of obs
            start_datetime = datetime.fromisoformat(start_time[:-7] + '+00:00')
            end_datetime = datetime.fromisoformat(end_time[:-7]+ '+00:00')
            num_days = (end_datetime - start_datetime).days
            if asset_df.shape[0] < (num_days-1):
                print(f"Did not obtain expected number of days for {asset} {metric}; there were {num_days-asset_df.shape[0]} fewer days than expected.")

            # Add data to master dataframe
            asset_results_df = asset_results_df.merge(asset_df, 
                                                        on=['asset', 'time'], 
                                                        how='outer', 
                                                        validate='one_to_one')
            
            # Space out the calls
            time.sleep(5)

        # Add this asset's results to the overall df
        results_df = pd.concat((results_df, asset_results_df))

        # Space out the calls across assets
        time.sleep(10)

    return results_df


def pullFirstTradeDates(base_url: str, base_params: dict, 
    cm_asset_universe: list, legit_exchanges: list, num_retries: int=3) -> pd.DataFrame:
    """
    Get the date of the first tradable market date for each asset in the given asset universe on legitimate exchanges.

    Parameters:
    -----------
    base_url: str
        The base URL of the API endpoint to use.

    base_params: dict
        A dictionary containing the parameters to include in the API calls.

    cm_asset_universe: list
        A list of assets to find the first tradable market dates for.

    legit_exchanges: list
        A list of legitimate exchanges to include.

    num_retries: int, default=3
        The number of times to retry an API call if it fails.

    Returns:
    --------
    pandas.DataFrame
        A DataFrame containing the date of the first tradable market for each asset in the given asset universe on legitimate exchanges.
    """
    # initialize dataframe to track these data with manually calc'ed dates for btc eth and usdt
    asset_ft_df = pd.DataFrame(data={'asset': ['btc', 'eth', 'usdt'],
                                    'date_first_tradable': ['2013-01-14', '2015-08-07', '2015-02-01']})

    # initiate params for coming calls
    params = base_params.copy()
    endpoint = "catalog-all/markets"

    for i in range(len(cm_asset_universe)):
        # Update asset
        asset = cm_asset_universe[i]

        # Monitor progress
        print(f"Processing the {i+1}th asset ({(i+1)/len(cm_asset_universe)*100:.2f}%): {asset}")

        # skip if btc, eth, or usdt as too many markets; i did manually
        if asset in ['btc', 'eth', 'usdt']:
            continue

        # determine the markets for this asset
        asset_markets = asset_metrics_df[asset_metrics_df.asset==asset].markets.values[0]

        # determine relevant markets that are spot and on relevant exchanges for this asset
        rel_asset_markets = [m for m in asset_markets if (m.split("-")[0] in legit_exchanges) & ('spot' in m)]

        # obtain the market info for all of these markets
        if len(rel_asset_markets) > 0:
            # obtain the market info
            params['markets'] = rel_asset_markets
            response = makeCmApiCall(base_url, endpoint, params)
            markets_info = response.json()['data']

            # calculate the minimum date across these markets
            min_date = datetime.utcnow().replace(tzinfo=pytz.utc)
            for market in markets_info:
                market_min_date = datetime.fromisoformat(market['min_time'][:-7] + '+00:00')
                if market_min_date.year < 2012: # some dates are just straight wrong so try to avoid them here
                    continue
                if market_min_date < min_date:
                    min_date = market_min_date
            
            # append this data
            asset_ft_df = pd.concat((asset_ft_df, pd.DataFrame(data={'asset': [asset], 
                                                                    'date_first_tradable': [min_date.strftime('%Y-%m-%d')]})))
                
            # space out the calls
            time.sleep(5)
        else:
            continue

    # return
    return asset_ft_df.sort_values(by='asset', ignore_index=True)



if __name__ == "__main__":
    # set parameters
    cmc_assets_fp = "../data/raw/cmc_asset_universe.pkl"
    cm_api_fp = '../../admin/coinmetrics.txt'
    cw_fp = "../data/raw/coinmetrics_cmc_cw.pkl"
    cm_asset_first_tradable_fp = '../data/raw/coinmetrics_assets_first_tradable.pkl'
    panel_fp = "../data/raw/coinmetrics_initial_panel.pkl"
    base_url = 'https://api.coinmetrics.io/v4/'
    cm_assets_wo_markets = ['atb', 'bcy', 'bits_bitswift', 'botx', 'bps', 'btcd', 'btcp', 'bz', 'drop', 'dtb', 'gam', 'golos', 'inb',
        'ino', 'ipc', 'lky', 'lmc', 'lrg', 'mcap', 'msp', 'net_nimiqexchangetoken', 'nxm', 'pnt_pnetwork', 'qau', 'qbit', 'rhoc', 'sjcx', 
        'thr', 'tio', 'tnc_tnccoin', 'usnbt', 'vrm', 'xbc', 'yoc']
    cm_assets_wo_markets.extend(['1st', '2give', 'aby', 'adk', 'adt', 'aeon', 'banca', 'bay', 'bitb', 'bitcny',
        'blk', 'block', 'bmc', 'bnk', 'brc', 'brk', 'bscpad', 'bsd_bitsend', 'bta',
        'bto', 'btu', 'btx', 'c20', 'ccxx', 'cfi', 'cnx', 'con', 'cosm', 'cpx', 'crw',
        'cure', 'dad', 'dcn', 'dct', 'ddd', 'dope', 'dpy', 'drg', 'dtr', 'dvi',
        'eauric', 'eca', 'edr_endorprotocol', 'egc', 'elama', 'emc', 'eosdac', 'etz',
        'excl', 'exp', 'fei', 'fldc', 'ftc', 'gard', 'geo', 'glc', 'gny', 'grc', 'hmq',
        'hmr', 'hns', 'hoge', 'hst', 'html', 'hyn', 'idh', 'incnt', 'ink', 'ioc',
        'ion', 'iop', 'ipx', 'jnt', 'kbc', 'krt', 'la', 'lbtc_lightningbitcoin',
        'lcc', 'maro', 'med', 'min', 'mix', 'mnx', 'morph', 'mue', 'music', 'mxm',
        'myst', 'nanj', 'next', 'nlc2', 'nlg', 'nvt', 'nxc', 'odn', 'ohm', 'ok', 'otn',
        'part', 'peak', 'pink', 'plbt', 'plf', 'plr', 'pltc', 'pot', 'ppp', 'ptoy',
        'qbt_qbao', 'qc', 'qrl', 'qwark', 'ret', 'rise', 'rox', 'rvr', 'safemoon',
        'sent', 'shift', 'sib', 'slr', 'sls', 'smart', 'soul_phantasma', 'sphtx',
        'ssx', 'stq', 'stx_stox', 'swt', 'swth', 'taas', 'taboo', 'ten', 'thc', 'tips',
        'tix', 'tnc', 'tpay', 'tshp', 'tx', 'ubq', 'ukg', 'upp', 'val', 'vitae', 'vrc',
        'xaur', 'xdn', 'xel', 'xmy', 'xsgd', 'xst', 'xtp', 'xwc', 'yffii', 'zcl', 'zlw',
        'zrc', 'zsc', 'zyn'])
    initial_asset_metrics = ['SplyAct1yr', 'SplyActEver', 'SplyCur', 'SplyFF', 
        'CapMrktCurUSD', 'CapMrktEstUSD', 'CapMrktFFUSD', 'CapRealUSD', 'ReferenceRateUSD']
    legit_exchanges = ['binance', 'binance.us', 'bitfinex', 'coinbase', 'crypto.com', 'ftx', 'ftx.us',
                   'gemini', 'huobi', 'kraken', 'kucoin', 'mt.gox', 'okex', 'poloniex', 'uniswap_v2_eth']
    
    # initialize client class
    with open(cm_api_fp) as f:
        API_KEY = f.readlines()
        API_KEY = API_KEY[0].strip()
    client = CoinMetricsClient(API_KEY)

    # set up base parameters
    base_params = {'api_key': API_KEY}

    # obtain coinmetrics assets
    cw_df = formCoinmetricsAssetUniverse(client, cmc_assets_fp)
    cw_df.to_pickle(cw_fp)
    cm_asset_universe = list(np.unique(cw_df[~cw_df.asset_cm.isnull()].asset_cm.values))
    cm_asset_universe = [a for a in cm_asset_universe if a not in cm_assets_wo_markets]

    # pull all metrics for assets in universe
    asset_metrics_df = pullAssetMetrics(client, base_url, base_params, cm_asset_universe, initial_asset_metrics)
    asset_metrics_df.to_pickle(panel_fp)

    # pull first date for assets
    asset_df = pullFirstTradeDates(base_url, base_params, cm_asset_universe, legit_exchanges)
    asset_df.to_pickle(cm_asset_first_tradable_fp)


# TODO save this df as the asset_df for just asset specific covariates


# TODO: ask coinmetrics if i can obtain access to the endpoints:
# -taxonomy/assets

# TODO: pull exchange metrics for those so maybe just volume?
# TODO: pull open interest in futures? maybe as macro variable but lets see if i can get for a bunch of assets
# TODO: pull the institutions and associated metrics to see if anything of interest
# TODO: pull bid and ask price for markets of target coins?
# TODO: pull defi balance sheets
# TODO: see if i am missing anything else to pull

all_metrics = []
for metrics in asset_metrics_df.metrics.values:
    if metrics is not np.nan:
        for metric in metrics:
            all_metrics.append(metric['metric'])
list(np.unique(np.array(all_metrics)))
