==> code/5a-pull_full_coinapi.ipynb <==
# TODO pull best ask and bid from those markets to obtain volume weighted spread

url = 'https://rest.coinapi.io/v1/quotes/BITSTAMP_SPOT_BTC_USD/history?time_start=2016-01-01T00:00:00'


time_start = 1
time_end = 2
limit = 50000

# TODO pull hourly data
